+++ CHANGELOG.md
@@ -3,6 +3,13 @@ This file is a running track of new features and fixes to each version of the pa
 
 This project follows [Semantic Versioning](http://semver.org) guidelines.
 
+## v0.6.0-pre.8 (Courageous Carniadactylus)
+### Fixed
+* `[pre.7]` — Fixes bug with subuser checkbox display.
+
+### Changed
+* Subuser permissions are now stored in `Permission::list()` to make views way cleaner and make adding to views significantly cleaner.
+
 ## v0.6.0-pre.7 (Courageous Carniadactylus)
 ### Fixed
 * `[pre.6]` — Addresses misconfigured console queue that was still sending data way to quickly thus causing the console to explode on some devices when large amounts of data were sent.